bugfix: earc failing with non-EFSOI experiment do to lack of EFSOI files

+++ jobs/rocoto/earc.sh
@@ -54,13 +54,14 @@ if [ $status -ne 0 ]; then
    exit $status
 fi
 
-$HOMEgfs/ush/hpssarch_gen.sh efsoigdas 
-status=$?
-if [ $status -ne 0 ]; then
-   echo "$HOMEgfs/ush/hpssarch_gen_EFSOI.sh enkf${CDUMP} failed, ABORT!"
-   exit $status
+if [ $DO_EFSOI = "YES" ]; then
+   $HOMEgfs/ush/hpssarch_gen.sh efsoigdas 
+   status=$?
+   if [ $status -ne 0 ]; then
+      echo "$HOMEgfs/ush/hpssarch_gen_EFSOI.sh enkf${CDUMP} failed, ABORT!"
+      exit $status
+   fi
 fi
-
 
 cd $ROTDIR
 
@@ -117,31 +118,32 @@ if [[ $ENSGRP -gt 0 ]] && [[ $HPSSARCH = "YES" ]]; then
        fi
      fi
 
-     htar -P -cvf $ATARDIR/$CDATE/efsoi${CDUMP}_grp${ENSGRP}.tar `cat $ARCH_LIST/efsoi${CDUMP}_grp${n}.txt`
-     status=$?
-     if [ $status -ne 0  -a $CDATE -ge $firstday ]; then
-         echo "HTAR $CDATE efsoi${CDUMP}_grp${ENSGRP}.tar failed"
-         exit $status
-     fi
-
-     if [ $SAVEWARMICA = "YES" -a $cyc -eq $EARCINC_CYC ]; then
-       htar -P -cvf $ATARDIR/$CDATE/efsoi${CDUMP}_restarta_grp${ENSGRP}.tar `cat $ARCH_LIST/efsoi${CDUMP}_restarta_grp${n}.txt`
+     if [ $DO_EFSOI = "YES" ]; then
+       htar -P -cvf $ATARDIR/$CDATE/efsoi${CDUMP}_grp${ENSGRP}.tar `cat $ARCH_LIST/efsoi${CDUMP}_grp${n}.txt`
        status=$?
-       if [ $status -ne 0 ]; then
-           echo "HTAR $CDATE efsoi${CDUMP}_restarta_grp${ENSGRP}.tar failed"
+       if [ $status -ne 0  -a $CDATE -ge $firstday ]; then
+           echo "HTAR $CDATE efsoi${CDUMP}_grp${ENSGRP}.tar failed"
            exit $status
        fi
-     fi
-
-     if [ $SAVEWARMICB = "YES"  -a $cyc -eq $EARCICS_CYC ]; then
-       htar -P -cvf $ATARDIR/$CDATE/efsoi${CDUMP}_restartb_grp${ENSGRP}.tar `cat $ARCH_LIST/efsoi${CDUMP}_restartb_grp${n}.txt`
-       status=$?
-       if [ $status -ne 0 ]; then
-           echo "HTAR $CDATE efsoi${CDUMP}_restartb_grp${ENSGRP}.tar failed"
-           exit $status
+  
+       if [ $SAVEWARMICA = "YES" -a $cyc -eq $EARCINC_CYC ]; then
+         htar -P -cvf $ATARDIR/$CDATE/efsoi${CDUMP}_restarta_grp${ENSGRP}.tar `cat $ARCH_LIST/efsoi${CDUMP}_restarta_grp${n}.txt`
+         status=$?
+         if [ $status -ne 0 ]; then
+             echo "HTAR $CDATE efsoi${CDUMP}_restarta_grp${ENSGRP}.tar failed"
+             exit $status
+         fi
        fi
-     fi
-
+  
+       if [ $SAVEWARMICB = "YES"  -a $cyc -eq $EARCICS_CYC ]; then
+         htar -P -cvf $ATARDIR/$CDATE/efsoi${CDUMP}_restartb_grp${ENSGRP}.tar `cat $ARCH_LIST/efsoi${CDUMP}_restartb_grp${n}.txt`
+         status=$?
+         if [ $status -ne 0 ]; then
+             echo "HTAR $CDATE efsoi${CDUMP}_restartb_grp${ENSGRP}.tar failed"
+             exit $status
+         fi
+       fi
+     fi # $DO_EFSOI = "YES" 
 
    fi # CDATE>SDATE
 
@@ -161,12 +163,14 @@ if [ $ENSGRP -eq 0 ]; then
             exit $status
         fi
 
-        htar -P -cvf $ATARDIR/$CDATE/efsoi${CDUMP}.tar `cat $ARCH_LIST/efsoi${CDUMP}.txt`
-        status=$?
-        if [ $status -ne 0  -a $CDATE -ge $firstday ]; then
-            echo "HTAR $CDATE efsoi${CDUMP}.tar failed"
-            exit $status
-        fi
+        if [ $DO_EFSOI = "YES" ]; then
+           htar -P -cvf $ATARDIR/$CDATE/efsoi${CDUMP}.tar `cat $ARCH_LIST/efsoi${CDUMP}.txt`
+           status=$?
+           if [ $status -ne 0  -a $CDATE -ge $firstday ]; then
+               echo "HTAR $CDATE efsoi${CDUMP}.tar failed"
+               exit $status
+           fi
+        fi # $DO_EFSOI = "YES"
 
     fi
 
@@ -178,19 +182,13 @@ if [ $ENSGRP -eq 0 ]; then
     $NCP $ROTDIR/enkf${CDUMP}.$PDY/$cyc/$COMPONENT/${CDUMP}.t${cyc}z.gsistat.ensmean  gsistat.${CDUMP}.${CDATE}.ensmean
 
     if [ $CDUMP_ENKF != "GDAS" ]; then
-		$NCP $ROTDIR/enkfgfs.$PDY/$cyc/$COMPONENT/${CDUMP}.t${cyc}z.enkfstat         enkfstat.gfs.$CDATE
-		$NCP $ROTDIR/enkfgfs.$PDY/$cyc/$COMPONENT/${CDUMP}.t${cyc}z.gsistat.ensmean  gsistat.gfs.${CDATE}.ensmean
-	fi
+        $NCP $ROTDIR/enkfgfs.$PDY/$cyc/$COMPONENT/${CDUMP}.t${cyc}z.enkfstat         enkfstat.gfs.$CDATE
+	$NCP $ROTDIR/enkfgfs.$PDY/$cyc/$COMPONENT/${CDUMP}.t${cyc}z.gsistat.ensmean  gsistat.gfs.${CDATE}.ensmean
+    fi
 
-
-    $NCP $ROTDIR/efsoi${CDUMP}.$PDY/$cyc/$COMPONENT/${CDUMP}.t${cyc}z.enkfstat         efsoistat.${CDUMP}.$CDATE
-    $NCP $ROTDIR/efsoi${CDUMP}.$PDY/$cyc/$COMPONENT/${CDUMP}.t${cyc}z.gsistat.ensmean  efsoigsistat.${CDUMP}.${CDATE}.ensmean
-
-    if [ $CDUMP_ENKF != "GDAS" ]; then
-		$NCP $ROTDIR/efsoigfs.$PDY/$cyc/$COMPONENT/${CDUMP}.t${cyc}z.enkfstat         efsoistat.gfs.$CDATE
-		$NCP $ROTDIR/efsoigfs.$PDY/$cyc/$COMPONENT/${CDUMP}.t${cyc}z.gsistat.ensmean  efsoigsistat.gfs.${CDATE}.ensmean
-	fi
-
+    if [ $DO_EFSOI = "YES" ]; then
+        $NCP $ROTDIR/efsoi${CDUMP}.$PDY/$cyc/$COMPONENT/${CDUMP}.t${cyc}z.enkfstat         efsoistat.${CDUMP}.$CDATE
+    fi
 
 fi
 
@@ -236,46 +234,49 @@ if [ $ENSGRP -eq 0 ]; then
 	    fi
 	done
 
-        COMIN_ENS="$ROTDIR/efsoigdas.$gPDY/$gcyc/$COMPONENT"
-	if [ -d $COMIN_ENS ] ; then
-	    rm -rf $COMIN_ENS/*f012*nc
-	    rm -rf $COMIN_ENS/*f018*nc
-            for imem in $(seq 1 $NMEM_ENKF); do
-                memchar="mem"$(printf %03i $imem)
-	        for file in `ls $COMIN_ENS/$memchar |grep -v atmf024`; do
-                    rm -rf $COMIN_ENS/$memchar/$file
+       if [ $DO_EFSOI = "YES" ]; then
+            COMIN_ENS="$ROTDIR/efsoigdas.$gPDY/$gcyc/$COMPONENT"
+            if [ -d $COMIN_ENS ] ; then
+                rm -rf $COMIN_ENS/*f012*nc
+                rm -rf $COMIN_ENS/*f018*nc
+                for imem in $(seq 1 $NMEM_ENKF); do
+                    memchar="mem"$(printf %03i $imem)
+                    for file in `ls $COMIN_ENS/$memchar |grep -v atmf024`; do
+                       rm -rf $COMIN_ENS/$memchar/$file
+                    done
                 done
-            done
-	fi
+   	    fi
+       fi #  $DO_EFSOI = "YES"
+   
+       # Advance to next cycle
+       GDATE=$($NDATE +$assim_freq $GDATE)
 
-	# Advance to next cycle
-	GDATE=$($NDATE +$assim_freq $GDATE)
+    done 
 
-    done
-
-    # Now do EFSOI - needs to be kept around longer
-    # Start start and end dates to remove
-    GDATEEND=$($NDATE -${RMOLDEND_EFSOI:-36}  $CDATE)
-    GDATE=$($NDATE -${RMOLDSTD_ENKF:-120} $CDATE)
-    while [ $GDATE -le $GDATEEND ]; do
-
-	gPDY=$(echo $GDATE | cut -c1-8)
-	gcyc=$(echo $GDATE | cut -c9-10)
-
-        COMIN_ENS="$ROTDIR/efsoigdas.$gPDY/$gcyc/$COMPONENT"
-	[[ -d $COMIN_ENS ]] && rm -rf $COMIN_ENS
-
-	# Remove any empty directories
-	COMIN_ENS="$ROTDIR/efsoigdas.$gPDY/$COMPONENT"
-	if [ -d $COMIN_ENS ] ; then
-	    [[ ! "$(ls -A $COMIN_ENS)" ]] && rm -rf $COMIN_ENS
-	fi
- 
-	# Advance to next cycle
-	GDATE=$($NDATE +$assim_freq $GDATE)
-
-    done
-
+    if [ $DO_EFSOI = "YES" ]; then
+        # Now do EFSOI - needs to be kept around longer
+        # Start start and end dates to remove
+        GDATEEND=$($NDATE -${RMOLDEND_EFSOI:-36}  $CDATE)
+        GDATE=$($NDATE -${RMOLDSTD_ENKF:-120} $CDATE)
+        while [ $GDATE -le $GDATEEND ]; do
+    
+           gPDY=$(echo $GDATE | cut -c1-8)
+           gcyc=$(echo $GDATE | cut -c9-10)
+    
+           COMIN_ENS="$ROTDIR/efsoigdas.$gPDY/$gcyc/$COMPONENT"
+             [[ -d $COMIN_ENS ]] && rm -rf $COMIN_ENS
+    
+           # Remove any empty directories
+           COMIN_ENS="$ROTDIR/efsoigdas.$gPDY/$COMPONENT"
+           if [ -d $COMIN_ENS ] ; then
+               [[ ! "$(ls -A $COMIN_ENS)" ]] && rm -rf $COMIN_ENS
+           fi
+           
+           # Advance to next cycle
+           GDATE=$($NDATE +$assim_freq $GDATE)
+           
+       done
+    fi # $DO_EFSOI = "YES" 
 
 fi
